+++ api/typescript/src/core/kurtosis_core_rpc_api_bindings/connect/api_container_service_pb.js
@@ -41,6 +41,17 @@ export const KurtosisFeatureFlag = proto3.makeEnum(
 );
 
 /**
+ * @generated from enum api_container_api.RestartPolicy
+ */
+export const RestartPolicy = proto3.makeEnum(
+  "api_container_api.RestartPolicy",
+  [
+    {no: 0, name: "NEVER"},
+    {no: 1, name: "ALWAYS"},
+  ],
+);
+
+/**
  * ==============================================================================================
  *                           Shared Objects (Used By Multiple Endpoints)
  * ==============================================================================================
@@ -590,5 +601,22 @@ export const ConnectServicesArgs = proto3.makeMessageType(
 export const ConnectServicesResponse = proto3.makeMessageType(
   "api_container_api.ConnectServicesResponse",
   [],
+);
+
+/**
+ * @generated from message api_container_api.GetStarlarkRunResponse
+ */
+export const GetStarlarkRunResponse = proto3.makeMessageType(
+  "api_container_api.GetStarlarkRunResponse",
+  () => [
+    { no: 1, name: "package_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 2, name: "serialized_script", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 3, name: "serialized_params", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 4, name: "parallelism", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
+    { no: 5, name: "relative_path_to_main_file", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 6, name: "main_function_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 7, name: "experimental_features", kind: "enum", T: proto3.getEnumType(KurtosisFeatureFlag), repeated: true },
+    { no: 8, name: "restart_policy", kind: "enum", T: proto3.getEnumType(RestartPolicy) },
+  ],
 );