+++ core/server/api_container/server/api_container_service.go
@@ -87,6 +87,10 @@ type ApiContainerService struct {
 
 	startosisModuleContentProvider startosis_packages.PackageContentProvider
 
+	restartPolicy kurtosis_core_rpc_api_bindings.RestartPolicy
+
+	starlarkRun *kurtosis_core_rpc_api_bindings.GetStarlarkRunResponse
+
 	metricsClient metrics_client.MetricsClient
 }
 
@@ -95,6 +99,7 @@ func NewApiContainerService(
 	serviceNetwork service_network.ServiceNetwork,
 	startosisRunner *startosis_engine.StartosisRunner,
 	startosisModuleContentProvider startosis_packages.PackageContentProvider,
+	restartPolicy kurtosis_core_rpc_api_bindings.RestartPolicy,
 	metricsClient metrics_client.MetricsClient,
 ) (*ApiContainerService, error) {
 	service := &ApiContainerService{
@@ -102,18 +107,30 @@ func NewApiContainerService(
 		serviceNetwork:                 serviceNetwork,
 		startosisRunner:                startosisRunner,
 		startosisModuleContentProvider: startosisModuleContentProvider,
-		metricsClient:                  metricsClient,
+		restartPolicy:                  restartPolicy,
+		starlarkRun: &kurtosis_core_rpc_api_bindings.GetStarlarkRunResponse{
+			PackageId:              startosis_constants.PackageIdPlaceholderForStandaloneScript,
+			SerializedScript:       "",
+			SerializedParams:       "",
+			Parallelism:            0,
+			RelativePathToMainFile: startosis_constants.PlaceHolderMainFileForPlaceStandAloneScript,
+			MainFunctionName:       "",
+			ExperimentalFeatures:   []kurtosis_core_rpc_api_bindings.KurtosisFeatureFlag{},
+			RestartPolicy:          kurtosis_core_rpc_api_bindings.RestartPolicy_NEVER,
+		},
+		metricsClient: metricsClient,
 	}
 
 	return service, nil
 }
 
-func (apicService ApiContainerService) RunStarlarkScript(args *kurtosis_core_rpc_api_bindings.RunStarlarkScriptArgs, stream kurtosis_core_rpc_api_bindings.ApiContainerService_RunStarlarkScriptServer) error {
+func (apicService *ApiContainerService) RunStarlarkScript(args *kurtosis_core_rpc_api_bindings.RunStarlarkScriptArgs, stream kurtosis_core_rpc_api_bindings.ApiContainerService_RunStarlarkScriptServer) error {
 	serializedStarlarkScript := args.GetSerializedScript()
 	serializedParams := args.GetSerializedParams()
 	parallelism := int(args.GetParallelism())
 	dryRun := shared_utils.GetOrDefaultBool(args.DryRun, defaultStartosisDryRun)
 	mainFuncName := args.GetMainFunctionName()
+	experimentalFeatures := args.GetExperimentalFeatures()
 	cloudUserId := shared_utils.GetOrDefaultString(args.CloudUserId, defaultCloudUserId)
 	cloudInstanceID := shared_utils.GetOrDefaultString(args.CloudInstanceId, defaultCloudInstanceId)
 
@@ -121,14 +138,25 @@ func (apicService ApiContainerService) RunStarlarkScript(args *kurtosis_core_rpc
 	if metricsErr != nil {
 		logrus.Warn("An error occurred tracking kurtosis run event")
 	}
+	noPackageReplaceOptions := map[string]string{}
 
-	nopackageReplaceOptions := map[string]string{}
+	apicService.runStarlark(parallelism, dryRun, startosis_constants.PackageIdPlaceholderForStandaloneScript, noPackageReplaceOptions, mainFuncName, startosis_constants.PlaceHolderMainFileForPlaceStandAloneScript, serializedStarlarkScript, serializedParams, args.GetExperimentalFeatures(), stream)
 
-	apicService.runStarlark(parallelism, dryRun, startosis_constants.PackageIdPlaceholderForStandaloneScript, nopackageReplaceOptions, mainFuncName, startosis_constants.PlaceHolderMainFileForPlaceStandAloneScript, serializedStarlarkScript, serializedParams, args.GetExperimentalFeatures(), stream)
+	apicService.starlarkRun = &kurtosis_core_rpc_api_bindings.GetStarlarkRunResponse{
+		PackageId:              startosis_constants.PackageIdPlaceholderForStandaloneScript,
+		SerializedScript:       serializedStarlarkScript,
+		SerializedParams:       serializedParams,
+		Parallelism:            int32(parallelism),
+		RelativePathToMainFile: startosis_constants.PlaceHolderMainFileForPlaceStandAloneScript,
+		MainFunctionName:       mainFuncName,
+		ExperimentalFeatures:   experimentalFeatures,
+		RestartPolicy:          apicService.restartPolicy,
+	}
+
 	return nil
 }
 
-func (apicService ApiContainerService) UploadStarlarkPackage(server kurtosis_core_rpc_api_bindings.ApiContainerService_UploadStarlarkPackageServer) error {
+func (apicService *ApiContainerService) UploadStarlarkPackage(server kurtosis_core_rpc_api_bindings.ApiContainerService_UploadStarlarkPackageServer) error {
 	var packageId string
 	serverStream := grpc_file_streaming.NewServerStream[kurtosis_core_rpc_api_bindings.StreamedDataChunk, emptypb.Empty](server)
 
@@ -162,7 +190,7 @@ func (apicService ApiContainerService) UploadStarlarkPackage(server kurtosis_cor
 	return nil
 }
 
-func (apicService ApiContainerService) InspectFilesArtifactContents(_ context.Context, args *kurtosis_core_rpc_api_bindings.InspectFilesArtifactContentsRequest) (*kurtosis_core_rpc_api_bindings.InspectFilesArtifactContentsResponse, error) {
+func (apicService *ApiContainerService) InspectFilesArtifactContents(_ context.Context, args *kurtosis_core_rpc_api_bindings.InspectFilesArtifactContentsRequest) (*kurtosis_core_rpc_api_bindings.InspectFilesArtifactContentsResponse, error) {
 	artifactIdentifier := ""
 	if args.GetFileNamesAndUuid().GetFileUuid() != emptyFileArtifactIdentifier {
 		artifactIdentifier = args.GetFileNamesAndUuid().GetFileUuid()
@@ -192,7 +220,7 @@ func (apicService ApiContainerService) InspectFilesArtifactContents(_ context.Co
 	}, nil
 }
 
-func (apicService ApiContainerService) RunStarlarkPackage(args *kurtosis_core_rpc_api_bindings.RunStarlarkPackageArgs, stream kurtosis_core_rpc_api_bindings.ApiContainerService_RunStarlarkPackageServer) error {
+func (apicService *ApiContainerService) RunStarlarkPackage(args *kurtosis_core_rpc_api_bindings.RunStarlarkPackageArgs, stream kurtosis_core_rpc_api_bindings.ApiContainerService_RunStarlarkPackageServer) error {
 	packageId := args.GetPackageId()
 	parallelism := int(args.GetParallelism())
 	dryRun := shared_utils.GetOrDefaultBool(args.DryRun, defaultStartosisDryRun)
@@ -236,10 +264,21 @@ func (apicService ApiContainerService) RunStarlarkPackage(args *kurtosis_core_rp
 		logrus.Warn("An error occurred tracking kurtosis run event")
 	}
 	apicService.runStarlark(parallelism, dryRun, packageName, packageReplaceOptions, mainFuncName, relativePathToMainFile, scriptWithRunFunction, serializedParams, args.ExperimentalFeatures, stream)
+
+	apicService.starlarkRun = &kurtosis_core_rpc_api_bindings.GetStarlarkRunResponse{
+		PackageId:              packageId,
+		SerializedScript:       scriptWithRunFunction,
+		SerializedParams:       serializedParams,
+		Parallelism:            int32(parallelism),
+		RelativePathToMainFile: relativePathToMainFile,
+		MainFunctionName:       mainFuncName,
+		ExperimentalFeatures:   args.ExperimentalFeatures,
+		RestartPolicy:          apicService.restartPolicy,
+	}
 	return nil
 }
 
-func (apicService ApiContainerService) ExecCommand(ctx context.Context, args *kurtosis_core_rpc_api_bindings.ExecCommandArgs) (*kurtosis_core_rpc_api_bindings.ExecCommandResponse, error) {
+func (apicService *ApiContainerService) ExecCommand(ctx context.Context, args *kurtosis_core_rpc_api_bindings.ExecCommandArgs) (*kurtosis_core_rpc_api_bindings.ExecCommandResponse, error) {
 	serviceIdentifier := args.ServiceIdentifier
 	command := args.CommandArgs
 	execResult, err := apicService.serviceNetwork.RunExec(ctx, serviceIdentifier, command)
@@ -266,7 +305,7 @@ func (apicService ApiContainerService) ExecCommand(ctx context.Context, args *ku
 	return resp, nil
 }
 
-func (apicService ApiContainerService) WaitForHttpGetEndpointAvailability(ctx context.Context, args *kurtosis_core_rpc_api_bindings.WaitForHttpGetEndpointAvailabilityArgs) (*emptypb.Empty, error) {
+func (apicService *ApiContainerService) WaitForHttpGetEndpointAvailability(ctx context.Context, args *kurtosis_core_rpc_api_bindings.WaitForHttpGetEndpointAvailabilityArgs) (*emptypb.Empty, error) {
 
 	serviceIdentifier := args.ServiceIdentifier
 
@@ -291,7 +330,7 @@ func (apicService ApiContainerService) WaitForHttpGetEndpointAvailability(ctx co
 	return &emptypb.Empty{}, nil
 }
 
-func (apicService ApiContainerService) WaitForHttpPostEndpointAvailability(ctx context.Context, args *kurtosis_core_rpc_api_bindings.WaitForHttpPostEndpointAvailabilityArgs) (*emptypb.Empty, error) {
+func (apicService *ApiContainerService) WaitForHttpPostEndpointAvailability(ctx context.Context, args *kurtosis_core_rpc_api_bindings.WaitForHttpPostEndpointAvailabilityArgs) (*emptypb.Empty, error) {
 	serviceIdentifier := args.ServiceIdentifier
 
 	if err := apicService.waitForEndpointAvailability(
@@ -315,7 +354,7 @@ func (apicService ApiContainerService) WaitForHttpPostEndpointAvailability(ctx c
 	return &emptypb.Empty{}, nil
 }
 
-func (apicService ApiContainerService) GetServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.GetServicesArgs) (*kurtosis_core_rpc_api_bindings.GetServicesResponse, error) {
+func (apicService *ApiContainerService) GetServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.GetServicesArgs) (*kurtosis_core_rpc_api_bindings.GetServicesResponse, error) {
 	serviceInfos := map[string]*kurtosis_core_rpc_api_bindings.ServiceInfo{}
 	filterServiceIdentifiers := args.ServiceIdentifiers
 
@@ -347,12 +386,12 @@ func (apicService ApiContainerService) GetServices(ctx context.Context, args *ku
 	return resp, nil
 }
 
-func (apicService ApiContainerService) ConnectServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.ConnectServicesArgs) (*kurtosis_core_rpc_api_bindings.ConnectServicesResponse, error) {
+func (apicService *ApiContainerService) ConnectServices(_ context.Context, _ *kurtosis_core_rpc_api_bindings.ConnectServicesArgs) (*kurtosis_core_rpc_api_bindings.ConnectServicesResponse, error) {
 	resp := binding_constructors.NewConnectServicesResponse()
 	return resp, nil
 }
 
-func (apicService ApiContainerService) GetExistingAndHistoricalServiceIdentifiers(_ context.Context, _ *emptypb.Empty) (*kurtosis_core_rpc_api_bindings.GetExistingAndHistoricalServiceIdentifiersResponse, error) {
+func (apicService *ApiContainerService) GetExistingAndHistoricalServiceIdentifiers(_ context.Context, _ *emptypb.Empty) (*kurtosis_core_rpc_api_bindings.GetExistingAndHistoricalServiceIdentifiersResponse, error) {
 	allIdentifiers, err := apicService.serviceNetwork.GetExistingAndHistoricalServiceIdentifiers()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting existing and historical service identifiers")
@@ -374,7 +413,7 @@ func (apicService ApiContainerService) GetExistingAndHistoricalServiceIdentifier
 	return existingAndHistoricalServiceIdentifiersResponse, nil
 }
 
-func (apicService ApiContainerService) UploadFilesArtifact(server kurtosis_core_rpc_api_bindings.ApiContainerService_UploadFilesArtifactServer) error {
+func (apicService *ApiContainerService) UploadFilesArtifact(server kurtosis_core_rpc_api_bindings.ApiContainerService_UploadFilesArtifactServer) error {
 	var maybeArtifactName string
 	serverStream := grpc_file_streaming.NewServerStream[kurtosis_core_rpc_api_bindings.StreamedDataChunk, kurtosis_core_rpc_api_bindings.UploadFilesArtifactResponse](server)
 
@@ -410,7 +449,7 @@ func (apicService ApiContainerService) UploadFilesArtifact(server kurtosis_core_
 	return nil
 }
 
-func (apicService ApiContainerService) DownloadFilesArtifact(args *kurtosis_core_rpc_api_bindings.DownloadFilesArtifactArgs, server kurtosis_core_rpc_api_bindings.ApiContainerService_DownloadFilesArtifactServer) error {
+func (apicService *ApiContainerService) DownloadFilesArtifact(args *kurtosis_core_rpc_api_bindings.DownloadFilesArtifactArgs, server kurtosis_core_rpc_api_bindings.ApiContainerService_DownloadFilesArtifactServer) error {
 	artifactIdentifier := args.Identifier
 	if strings.TrimSpace(artifactIdentifier) == "" {
 		return stacktrace.NewError("Cannot download file with empty files artifact identifier")
@@ -465,7 +504,7 @@ func (apicService ApiContainerService) DownloadFilesArtifact(args *kurtosis_core
 	return nil
 }
 
-func (apicService ApiContainerService) StoreWebFilesArtifact(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StoreWebFilesArtifactArgs) (*kurtosis_core_rpc_api_bindings.StoreWebFilesArtifactResponse, error) {
+func (apicService *ApiContainerService) StoreWebFilesArtifact(_ context.Context, args *kurtosis_core_rpc_api_bindings.StoreWebFilesArtifactArgs) (*kurtosis_core_rpc_api_bindings.StoreWebFilesArtifactResponse, error) {
 	url := args.Url
 	artifactName := args.Name
 
@@ -487,7 +526,7 @@ func (apicService ApiContainerService) StoreWebFilesArtifact(ctx context.Context
 	return response, nil
 }
 
-func (apicService ApiContainerService) StoreFilesArtifactFromService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceArgs) (*kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceResponse, error) {
+func (apicService *ApiContainerService) StoreFilesArtifactFromService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceArgs) (*kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceResponse, error) {
 	serviceIdentifier := args.ServiceIdentifier
 	srcPath := args.SourcePath
 	name := args.Name
@@ -501,7 +540,7 @@ func (apicService ApiContainerService) StoreFilesArtifactFromService(ctx context
 	return response, nil
 }
 
-func (apicService ApiContainerService) ListFilesArtifactNamesAndUuids(_ context.Context, _ *emptypb.Empty) (*kurtosis_core_rpc_api_bindings.ListFilesArtifactNamesAndUuidsResponse, error) {
+func (apicService *ApiContainerService) ListFilesArtifactNamesAndUuids(_ context.Context, _ *emptypb.Empty) (*kurtosis_core_rpc_api_bindings.ListFilesArtifactNamesAndUuidsResponse, error) {
 	filesArtifactsNamesAndUuids := apicService.filesArtifactStore.GetFileNamesAndUuids()
 	var filesArtifactNamesAndUuids []*kurtosis_core_rpc_api_bindings.FilesArtifactNameAndUuid
 	for _, nameAndUuid := range filesArtifactsNamesAndUuids {
@@ -514,6 +553,10 @@ func (apicService ApiContainerService) ListFilesArtifactNamesAndUuids(_ context.
 	return &kurtosis_core_rpc_api_bindings.ListFilesArtifactNamesAndUuidsResponse{FileNamesAndUuids: filesArtifactNamesAndUuids}, nil
 }
 
+func (apicService *ApiContainerService) GetStarlarkRun(_ context.Context, _ *emptypb.Empty) (*kurtosis_core_rpc_api_bindings.GetStarlarkRunResponse, error) {
+	return apicService.starlarkRun, nil
+}
+
 // ====================================================================================================
 //
 //	Private helper methods
@@ -523,7 +566,7 @@ func transformPortSpecToApiPort(port *port_spec.PortSpec) (*kurtosis_core_rpc_ap
 	portNumUint16 := port.GetNumber()
 	portSpecProto := port.GetTransportProtocol()
 
-	// Yes, this isn't the most efficient way to do this, but the map is tiny so it doesn't matter
+	// Yes, this isn't the most efficient way to do this, but the map is tiny, so it doesn't matter
 	var apiProto kurtosis_core_rpc_api_bindings.Port_TransportProtocol
 	foundApiProto := false
 	for mappedApiProto, mappedPortSpecProto := range apiContainerPortProtoToPortSpecPortProto {
@@ -563,7 +606,7 @@ func transformPortSpecMapToApiPortsMap(apiPorts map[string]*port_spec.PortSpec) 
 	return result, nil
 }
 
-func (apicService ApiContainerService) waitForEndpointAvailability(
+func (apicService *ApiContainerService) waitForEndpointAvailability(
 	ctx context.Context,
 	serviceIdStr string,
 	httpMethod string,
@@ -659,7 +702,7 @@ func makeHttpRequest(httpMethod string, url string, body string) (*http.Response
 	return resp, nil
 }
 
-func (apicService ApiContainerService) getServiceInfoForIdentifier(ctx context.Context, serviceIdentifier string) (*kurtosis_core_rpc_api_bindings.ServiceInfo, error) {
+func (apicService *ApiContainerService) getServiceInfoForIdentifier(ctx context.Context, serviceIdentifier string) (*kurtosis_core_rpc_api_bindings.ServiceInfo, error) {
 	serviceObj, err := apicService.serviceNetwork.GetService(
 		ctx,
 		serviceIdentifier,
@@ -674,7 +717,7 @@ func (apicService ApiContainerService) getServiceInfoForIdentifier(ctx context.C
 	return serviceInfo, nil
 }
 
-func (apicService ApiContainerService) runStarlarkPackageSetup(
+func (apicService *ApiContainerService) runStarlarkPackageSetup(
 	packageId string,
 	clonePackage bool,
 	moduleContentIfLocal []byte,
@@ -718,7 +761,7 @@ func (apicService ApiContainerService) runStarlarkPackageSetup(
 	return string(mainScriptToExecute), kurtosisYml, nil
 }
 
-func (apicService ApiContainerService) runStarlark(
+func (apicService *ApiContainerService) runStarlark(
 	parallelism int,
 	dryRun bool,
 	packageId string,